+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 
 ### Added
 
+- Fixed bug with Jupyter server appearing as project filepath
 - Initial plugin and integration with Tinymist LSP created
   from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)